Refactor DatabaseBlobStorageCleanupJob configuration
- Move BlobStorageConfiguration to the right package (was in autoconfigure from development when it started as an Autoconfiguration class)
- Move the @Bean creation out of DatabaseBlobStorageCleanupJob, make more sense to configure the whole system in one place (plus since it is not Autoconfiguration anymore we can't test on presence of DatabaseBlobStorage instance)
