+++ webapp/src/main/java/com/box/l10n/mojito/service/blobstorage/BlobStorageConfiguration.java
@@ -1,18 +1,24 @@
-package com.box.l10n.mojito.service.blobstorage.autoconfigure;
+package com.box.l10n.mojito.service.blobstorage;
 
 import com.amazonaws.services.s3.AmazonS3;
-import com.box.l10n.mojito.service.blobstorage.BlobStorage;
 import com.box.l10n.mojito.service.blobstorage.database.DatabaseBlobStorage;
 import com.box.l10n.mojito.service.blobstorage.database.DatabaseBlobStorageCleanupJob;
 import com.box.l10n.mojito.service.blobstorage.database.DatabaseBlobStorageConfigurationProperties;
 import com.box.l10n.mojito.service.blobstorage.database.MBlobRepository;
 import com.box.l10n.mojito.service.blobstorage.s3.S3BlobStorage;
 import com.box.l10n.mojito.service.blobstorage.s3.S3BlobStorageConfigurationProperties;
+import org.quartz.JobDetail;
+import org.quartz.SimpleTrigger;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
+import org.springframework.context.annotation.Profile;
+import org.springframework.scheduling.quartz.JobDetailFactoryBean;
+import org.springframework.scheduling.quartz.SimpleTriggerFactoryBean;
 
 /**
  * Configuration for {@link BlobStorage}
@@ -26,6 +32,8 @@ import org.springframework.context.annotation.Import;
 @Configuration
 public class BlobStorageConfiguration {
 
+    static Logger logger = LoggerFactory.getLogger(BlobStorageConfiguration.class);
+
     @ConditionalOnProperty(value = "l10n.blob-storage.type", havingValue = "s3")
     @Configuration
     static class S3BlobStorageConfigurationConfiguration {
@@ -38,12 +46,12 @@ public class BlobStorageConfiguration {
 
         @Bean
         public S3BlobStorage s3BlobStorage() {
+            logger.info("Configure S3BlobStorage");
             return new S3BlobStorage(amazonS3, s3BlobStorageConfigurationProperties);
         }
     }
 
     @ConditionalOnProperty(value = "l10n.blob-storage.type", havingValue = "database", matchIfMissing = true)
-    @Import(DatabaseBlobStorageCleanupJob.class)
     static class DatabaseBlobStorageConfiguration {
 
         @Autowired
@@ -54,7 +62,29 @@ public class BlobStorageConfiguration {
 
         @Bean
         public DatabaseBlobStorage databaseBlobStorage() {
+            logger.info("Configure DatabaseBlobStorage");
             return new DatabaseBlobStorage(databaseBlobStorageConfigurationProperties, mBlobRepository);
         }
+
+        @Bean(name = "jobDetailDatabaseBlobStorageCleanupJob")
+        public JobDetailFactoryBean jobDetailExpiringBlobCleanup() {
+            JobDetailFactoryBean jobDetailFactory = new JobDetailFactoryBean();
+            jobDetailFactory.setJobClass(DatabaseBlobStorageCleanupJob.class);
+            jobDetailFactory.setDescription("Cleanup expired blobs");
+            jobDetailFactory.setDurability(true);
+            return jobDetailFactory;
+        }
+
+        @Profile("!disablescheduling")
+        @Bean
+        public SimpleTriggerFactoryBean triggerExpiringBlobCleanup(@Qualifier("jobDetailDatabaseBlobStorageCleanupJob") JobDetail job) {
+            logger.info("Configure jobDetailDatabaseBlobStorageCleanupJob");
+            SimpleTriggerFactoryBean trigger = new SimpleTriggerFactoryBean();
+            trigger.setJobDetail(job);
+            trigger.setRepeatInterval(300000);
+            trigger.setRepeatCount(SimpleTrigger.REPEAT_INDEFINITELY);
+            return trigger;
+        }
     }
+
 }
